add ability of several users, small fixes

+++ main.js
@@ -8,22 +8,23 @@ import { Keyboard } from 'telegram-keyboard';
 // fs.writeFileSync(filename, data)
 
 const keyboard1 = Markup.inlineKeyboard([[
-		Markup.button.callback('Знаю', 'add_to_learned0'),
-		Markup.button.callback('Не знаю', 'add_to_need_to_repeat'),
-	],
-
-	[
-		Markup.button.callback('Озвучить', 'hear_word')
-	]
-
-
-])
-const keyboard2 = Markup.inlineKeyboard([[
-	Markup.button.callback('Я запомнил', 'add_to_learned1'),
+	Markup.button.callback('Знаю', 'add_to_learned0'),
+	Markup.button.callback('Не знаю', 'add_to_need_to_repeat'),
 ],
 
 [
 	Markup.button.callback('Озвучить', 'hear_word')
+]
+
+
+])
+const keyboard2 = Markup.inlineKeyboard([[
+	Markup.button.callback('Я запомнил', 'add_to_learned1'), Markup.button.callback('Не знаю', 'get_translation'),
+],
+
+[
+	Markup.button.callback('Озвучить', 'hear_word')
+
 ]
 
 
@@ -40,27 +41,30 @@ const keyboard3 = Markup.inlineKeyboard([[
 
 ])
 
-let current_word = {}
-let mode = ""
+let current = { mode: "", question: "" }
 
 let learn_new_words = (ctx) => {
-	mode = 'new'
+	current['mode'] = 'new'
 	let data = JSON.parse(fs.readFileSync("data/data.json"))
 	let user = { "learned": [], "need_to_repeat": [], "new": data.all.sort(() => Math.random() - 0.5) }
-	if (!fs.existsSync("user.json")) {
-		fs.writeFileSync("user.json", JSON.stringify(user))
+	if (!fs.existsSync(`user_configs/${ctx.chat.id}.json`)) {
+		fs.writeFileSync(`user_configs/${ctx.chat.id}.json`, JSON.stringify(user))
 	} else {
-		user = JSON.parse(fs.readFileSync("user.json"))
+		user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	}
-	let word = user.new.pop()
-	fs.writeFileSync("user.json", JSON.stringify(user))
-	ctx.reply(word, keyboard1)
+	if (user.new.length > 0) {
+		let word = user.new.pop()
+		ctx.reply(word, keyboard1)
+	} else {
+		ctx.reply('Вы ознакомились со всеми словами')
+	}
+
 }
 
 let check = async (ctx) => {
-	console.log(current_word)
 	let answ = ctx.message.text.toLocaleLowerCase()
 	let correct = 0
+	let current_word = current['question']['translations']
 	for (let i = 0; i < current_word.length; i++) {
 		let line = current_word[i]
 		for (let word of line) {
@@ -86,7 +90,7 @@ let check = async (ctx) => {
 	}
 
 	if (!correct) {
-		let answ = ""
+		let answ = current.question.word + " - "
 		for (let line = 0; line < current_word.length; line++) {
 			answ += current_word[line].join(', ')
 			answ += '\n\n'
@@ -99,16 +103,21 @@ let check = async (ctx) => {
 
 let repeat_words = (ctx) => {
 	let data = JSON.parse(fs.readFileSync("data/data.json"))
-	let user = JSON.parse(fs.readFileSync("user.json"))
+	let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	let need_to_repeat = user['need_to_repeat']
-	mode = 'repeating'
-	let word = need_to_repeat[Math.floor(Math.random()*need_to_repeat.length)]
-	current_word = data[word]
-	ctx.reply(word, keyboard2)
+	current['mode'] = 'repeating'
+	if (need_to_repeat.length > 0) {
+		let word = need_to_repeat[Math.floor(Math.random() * need_to_repeat.length)]
+		current['question'] = { translations: data[word], word: word }
+		ctx.reply(word, keyboard2)
+	} else {
+		ctx.reply('Вы выучили все указанные слова')
+	}
+
 }
 
 let statistics = (ctx) => {
-	let user = JSON.parse(fs.readFileSync("user.json"))
+	let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	let learned = user['learned'].length
 	let learning = user['need_to_repeat'].length
 	ctx.reply(`You know ${learned} words\nYou are learning ${learning} words`, keyboard3)
@@ -120,17 +129,32 @@ let actions = [learn_new_words, repeat_words]
 for (let i of [0, 1]) {
 	bot.action(`add_to_learned${i}`, (ctx) => {
 		let word = ctx.update.callback_query.message.text
-		let user = JSON.parse(fs.readFileSync("user.json"))
+		let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 		ctx.deleteMessage()
 		user.learned.push(word)
 		if (user.need_to_repeat.includes(word)) {
 			user.need_to_repeat.splice(user.need_to_repeat.indexOf(word), 1);
-	}
-		fs.writeFileSync("user.json", JSON.stringify(user))
+		}
+		if (user.new.includes(word)) {
+			user.new.pop()
+		}
+		fs.writeFileSync(`user_configs/${ctx.chat.id}.json`, JSON.stringify(user))
 		actions[i](ctx)
-	
+
 	})
 }
+
+bot.action(`get_translation`, async (ctx) => {
+	let current_word = current['question']['translations']
+	let answ = current.question.word + " - "
+	for (let line = 0; line < current_word.length; line++) {
+		answ += current_word[line].join(', ')
+		answ += '\n\n'
+	}
+	await ctx.reply(answ)
+	repeat_words(ctx)
+})
+
 
 bot.action(`hear_word`, async (ctx) => {
 	let word = ctx.update.callback_query.message.text
@@ -140,29 +164,41 @@ bot.action(`hear_word`, async (ctx) => {
 })
 
 bot.action(`learned_list`, async (ctx) => {
-	let user = JSON.parse(fs.readFileSync("user.json"))
+	let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	let tmp_file_name = Math.floor(Math.random() * 1000) + '.txt'
-	let data = user.learned.join('\n')
-	fs.writeFileSync(tmp_file_name, data)
-	await ctx.replyWithDocument({'source': fs.createReadStream(tmp_file_name)})
-	fs.unlinkSync(tmp_file_name)
+	if (user.learned.length > 0) {
+		let data = user.learned.join('\n')
+		fs.writeFileSync(tmp_file_name, data)
+		await ctx.replyWithDocument({ 'source': fs.createReadStream(tmp_file_name) })
+		fs.unlinkSync(tmp_file_name)
+	} else {
+		ctx.reply('Список пуст')
+	}
+
 })
 
 bot.action(`learning_list`, async (ctx) => {
-	let user = JSON.parse(fs.readFileSync("user.json"))
+	let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	let tmp_file_name = Math.floor(Math.random() * 1000) + '.txt'
-	let data = user.need_to_repeat.join('\n')
-	fs.writeFileSync(tmp_file_name, data)
-	await ctx.replyWithDocument({'source': fs.createReadStream(tmp_file_name)})
-	fs.unlinkSync(tmp_file_name)
+	if (user.need_to_repeat.length > 0) {
+		let data = user.need_to_repeat.join('\n')
+		fs.writeFileSync(tmp_file_name, data)
+		await ctx.replyWithDocument({ 'source': fs.createReadStream(tmp_file_name) })
+		fs.unlinkSync(tmp_file_name)
+	} else {
+		ctx.reply('Список пуст')
+	}
 })
 
 bot.action('add_to_need_to_repeat', (ctx) => {
 	ctx.deleteMessage()
 	let word = ctx.update.callback_query.message.text
-	let user = JSON.parse(fs.readFileSync("user.json"))
+	let user = JSON.parse(fs.readFileSync(`user_configs/${ctx.chat.id}.json`))
 	user.need_to_repeat.push(word)
-	fs.writeFileSync("user.json", JSON.stringify(user))
+	if (user.new.includes(word)) {
+		user.new.pop()
+	}
+	fs.writeFileSync(`user_configs/${ctx.chat.id}.json`, JSON.stringify(user))
 	learn_new_words(ctx)
 
 })
@@ -185,7 +221,7 @@ bot.on('text', async (ctx) => {
 			statistics(ctx)
 			break;
 		default:
-			if(mode === 'repeating'){
+			if (current['mode'] === 'repeating') {
 				check(ctx);
 				break;
 			}
